+++ src/Constraints/CreditCard.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace SLLH\IsoCodesValidator\Constraints;
+
+use Symfony\Component\Validator\Constraint;
+
+/**
+ * Class CreditCard.
+ *
+ * @Annotation
+ * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
+ *
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+class CreditCard extends Constraint
+{
+    public $message = 'This value is not a valid credit card scheme.';
+}