+++ src/Constraints/Isbn10.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace SLLH\IsoCodesValidator\Constraints;
+
+use Symfony\Component\Validator\Constraint;
+
+/**
+ * Class Isbn10.
+ *
+ * @Annotation
+ * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
+ *
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+class Isbn10 extends Constraint
+{
+    public $message = 'This value is not a valid ISBN-10.';
+}