+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * 1.1.0 (2015-??-??)
 
  * Deprecate `IsoCodesConstraintValidator` in favor of `AbstractIsoCodesConstraintValidator`.
+ * Add `CreditCard` constraint.
+ * Add `Ip` constraint.
+ * Add `Iban` constraint.
+ * Add `Isbn10` constraint.
 
 * 1.0.3 (2015-06-02)